don't delete file if there is no file

+++ Modules/IndividualAssessment/classes/class.ilIndividualAssessmentMemberGUI.php
@@ -212,7 +212,9 @@ class ilIndividualAssessmentMemberGUI extends AbstractCtrlAwareUploadHandler
 
         if ($grading->getFile() == '') {
             $storage = $this->getUserFileStorage();
-            $storage->deleteCurrentFile();
+            if (!$storage->isEmpty()) {
+                $storage->deleteCurrentFile();
+            }
         }
 
         if ($grading->isFinalized()) {
